veNTUre-X-Mercer-main/mercerchatbotbackend2/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
requests==2.32.3
python-multipart==0.0.9
google-cloud-storage==2.18.2
pdf2image==1.17.0
Pillow==10.4.0
pyngrok==7.2.0
pytesseract==0.3.10  # OCR support (requires system tesseract-ocr installed)
